Change wooid to string
The error message TypeError: expected str, bytes or os.PathLike object, not int indicates that there is an attempt to expand the user's home directory for an integer value.

Based on the code snippet you provided earlier, it seems like the issue is most likely in the put_request function defined in "apps/woocommerceconnector/woocommerceconnector/woocommerce_requests.py".

It's possible that the wooid variable passed to put_request is not a string, bytes, or os.PathLike object, but rather an integer. To confirm this, you can print the value of wooid before calling the put_request function. If it's an integer, you can try converting it to a string using the str() function before passing it to put_request.