Rework get_tax_account_head(tax)
In this version:

tax_title is assigned the value of tax.get("name") or tax.get("method_title"), and it's ensured to be a string.

The encode("utf-8") method is removed.

The SQL query string is constructed using the parameterized query approach. The value tax_title_str is used as a parameter, eliminating the need for manual encoding.

This should resolve the SQL syntax error you were encountering. Make sure to test this updated code to ensure it works as expected.